requirements.txt
streamlit
google-generativeai
youtube_transcript_api
fpdf
re
google-api-python-client
os
DateTime